feat(后台管理): 角色
1.角色基本代码
2.角色管理

+++ src/main/java/com/acme/acmemall/controller/admin/SystemAdminController.java
@@ -113,6 +113,9 @@ public class SystemAdminController extends ApiBase {
             return toResponsFail("参数错误");
         }
         Long roleId = reqObj.getLong("roleId");
+        if (roleId == null) {
+            return toResponsFail("参数错误");
+        }
         return sysRoleService.deleteRole(roleId);
     }
 }
